+++ GameLauncher/App/Classes/LauncherCore/Global/FunctionStatus.cs
@@ -1,6 +1,7 @@
 using GameLauncher.App.Classes.LauncherCore.FileReadWrite;
 using GameLauncher.App.Classes.LauncherCore.Lists;
 using GameLauncher.App.Classes.LauncherCore.Lists.JSON;
+using GameLauncher.App.Classes.LauncherCore.RPC;
 using GameLauncher.App.Classes.Logger;
 using GameLauncher.App.Classes.SystemPlatform.Linux;
 using GameLauncher.App.Classes.SystemPlatform.Windows;
@@ -13,6 +14,7 @@ using System.Globalization;
 using System.IO;
 using System.Management.Automation;
 using System.Net;
+using System.Threading;
 using System.Windows.Forms;
 
 namespace GameLauncher.App.Classes.LauncherCore.Global
@@ -57,6 +59,9 @@ namespace GameLauncher.App.Classes.LauncherCore.Global
     /* This is Used to call Certain Functions (Such as Completion Status or Function Callbacks) */
     class FunctionStatus
     {
+        /* Used to check if Launcher Launch had fully completed its checks */
+        public static bool LoadingComplete = false;
+
         /* Allows Registration Button to be Enabled/Disabled */
         public static bool AllowRegistration;
 
@@ -145,162 +150,6 @@ namespace GameLauncher.App.Classes.LauncherCore.Global
             else if (CurrentLang == "rus") return 121367723;
             else if (CurrentLang == "spa") return 101540466;
             else return 141805935;
-        }
-
-        public static void FirstTimeRun()
-        {
-            Log.Core("LAUNCHER: Checking InstallationDirectory: " + FileSettingsSave.GameInstallation);
-            if (string.IsNullOrEmpty(FileSettingsSave.GameInstallation))
-            {
-                Log.Core("LAUNCHER: First run!");
-
-                try
-                {
-                    Form welcome = new WelcomeScreen();
-                    DialogResult welcomereply = welcome.ShowDialog();
-
-                    if (welcomereply != DialogResult.OK)
-                    {
-                        Process.GetProcessById(Process.GetCurrentProcess().Id).Kill();
-                    }
-                    else
-                    {
-                        FileSettingsSave.CDN = SelectedCDN.CDNUrl;
-                        FileSettingsSave.SaveSettings();
-                    }
-                }
-                catch
-                {
-                    Log.Warning("LAUNCHER: CDN Source URL was Empty! Setting a Null Safe URL 'http://localhost'");
-                    FileSettingsSave.CDN = "http://localhost";
-                    Log.Core("LAUNCHER: Installation Directory was Empty! Creating and Setting Directory at " + AppDomain.CurrentDomain.BaseDirectory + "\\Game Files");
-                    FileSettingsSave.GameInstallation = AppDomain.CurrentDomain.BaseDirectory + "\\Game Files";
-                    FileSettingsSave.SaveSettings();
-                }
-
-                var fbd = new CommonOpenFileDialog
-                {
-                    EnsurePathExists = true,
-                    EnsureFileExists = false,
-                    AllowNonFileSystemItems = false,
-                    Title = "Select the location to Find or Download NFS:W",
-                    IsFolderPicker = true
-                };
-
-                if (fbd.ShowDialog() == CommonFileDialogResult.Ok)
-                {
-                    if (!HasWriteAccessToFolder(fbd.FileName))
-                    {
-                        Log.Error("LAUNCHER: Not enough permissions. Exiting.");
-                        MessageBox.Show(null, "You don't have enough permission to select this path as installation folder. Please select another directory.", "GameLauncher", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                        Environment.Exit(Environment.ExitCode);
-                    }
-
-                    if (fbd.FileName.Length == 3)
-                    {
-                        Log.Warning("LAUNCHER: Installing NFSW in root of the harddisk is not allowed.");
-                        MessageBox.Show(null, string.Format("Installing NFSW in root of the harddisk is not allowed. " +
-                            "Instead, we will install it on {0}.", AppDomain.CurrentDomain.BaseDirectory + "\\Game Files"), "GameLauncher", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                        FileSettingsSave.GameInstallation = AppDomain.CurrentDomain.BaseDirectory + "\\Game Files";
-                        FileSettingsSave.SaveSettings();
-                    }
-                    else if (fbd.FileName == AppDomain.CurrentDomain.BaseDirectory)
-                    {
-                        Directory.CreateDirectory("Game Files");
-                        Log.Warning("LAUNCHER: Installing NFSW in same directory where the launcher resides is disadvised.");
-                        MessageBox.Show(null, string.Format("Installing NFSW in same directory where the launcher resides is disadvised. " +
-                            "Instead, we will install it on {0}.", AppDomain.CurrentDomain.BaseDirectory + "\\Game Files"), "GameLauncher", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                        FileSettingsSave.GameInstallation = AppDomain.CurrentDomain.BaseDirectory + "\\Game Files";
-                        FileSettingsSave.SaveSettings();
-                    }
-                    else
-                    {
-                        Log.Core("LAUNCHER: Directory Set: " + fbd.FileName);
-                        FileSettingsSave.GameInstallation = fbd.FileName;
-                        FileSettingsSave.SaveSettings();
-                    }
-                }
-                else
-                {
-                    Log.Core("LAUNCHER: Exiting");
-                    Environment.Exit(Environment.ExitCode);
-                }
-                fbd.Dispose();
-            }
-
-            if (!DetectLinux.LinuxDetected())
-            {
-                switch (CheckFolder(FileSettingsSave.GameInstallation))
-                {
-                    case FolderType.IsSameAsLauncherFolder:
-                        Directory.CreateDirectory("Game Files");
-                        Log.Error("LAUNCHER: Installing NFSW in same location where the GameLauncher resides is NOT allowed.");
-                        MessageBox.Show(null, string.Format("Installing NFSW in same location where the GameLauncher resides is NOT allowed.\n" +
-                            "Instead, we will install it at {0}.", AppDomain.CurrentDomain.BaseDirectory + "Game Files"), "GameLauncher", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                        FileSettingsSave.GameInstallation = AppDomain.CurrentDomain.BaseDirectory + "\\Game Files";
-                        break;
-
-                    case FolderType.IsTempFolder:
-                    case FolderType.IsUsersFolders:
-                    case FolderType.IsProgramFilesFolder:
-                    case FolderType.IsWindowsFolder:
-                    case FolderType.IsRootFolder:
-                        String constructMsg = String.Empty;
-                        Directory.CreateDirectory("Game Files");
-                        constructMsg += "Using this location for Game Files is not allowed.\nThe following list are NOT allowed:\n\n";
-                        constructMsg += "• X:\\nfsw.exe (Root of Drive, such as C:\\ or D:\\, must be in a folder)\n";
-                        constructMsg += "• C:\\Program Files\n";
-                        constructMsg += "• C:\\Program Files (x86)\n";
-                        constructMsg += "• C:\\Users (Includes 'Desktop', 'Documents', 'Downloads')\n";
-                        constructMsg += "• C:\\Windows\n\n";
-                        constructMsg += "Instead, we will install the NFSW Game at " + AppDomain.CurrentDomain.BaseDirectory + "\\Game Files\n";
-
-                        MessageBox.Show(null, constructMsg, "GameLauncher", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                        Log.Error("LAUNCHER: Installing NFSW in a Restricted Location is not allowed.");
-                        FileSettingsSave.GameInstallation = AppDomain.CurrentDomain.BaseDirectory + "\\Game Files";
-                        break;
-                }
-                FileSettingsSave.SaveSettings();
-
-                /* Windows Defender (Windows 10) */
-                if (WindowsProductVersion.CachedWindowsNumber >= 10.0 && (FileSettingsSave.WindowsDefenderStatus == "Not Excluded" || FileSettingsSave.WindowsDefenderStatus == "Unknown"))
-                {
-                    Log.Core("WINDOWS DEFENDER: Windows 10 Detected! Running Exclusions for Core Folders");
-                    if (ManagementSearcher.SecurityCenter("AntivirusEnabled") == true && ManagementSearcher.SecurityCenter("AntispywareEnabled") == true)
-                    {
-                        /* Create Windows Defender Exclusion */
-                        try
-                        {
-                            Log.Info("WINDOWS DEFENDER: Excluding Core Folders");
-                            /* Add Exclusion to Windows Defender */
-                            using (PowerShell ps = PowerShell.Create())
-                            {
-                                ps.AddScript($"Add-MpPreference -ExclusionPath \"{AppDomain.CurrentDomain.BaseDirectory}\"");
-                                ps.AddScript($"Add-MpPreference -ExclusionPath \"{FileSettingsSave.GameInstallation}\"");
-                                var result = ps.Invoke();
-                            }
-
-                            FileSettingsSave.WindowsDefenderStatus = "Excluded";
-                            FileSettingsSave.SaveSettings();
-                        }
-                        catch (Exception ex)
-                        {
-                            Log.Error("WINDOWS DEFENDER: " + ex.Message);
-                            FileSettingsSave.WindowsDefenderStatus = "Not Excluded";
-                            FileSettingsSave.SaveSettings();
-                        }
-                    }
-                    else
-                    {
-                        FileSettingsSave.WindowsDefenderStatus = "Not Supported";
-                        FileSettingsSave.SaveSettings();
-                    }
-                }
-                else if (WindowsProductVersion.CachedWindowsNumber >= 10.0 && !string.IsNullOrEmpty(FileSettingsSave.WindowsDefenderStatus))
-                {
-                    Log.Core("WINDOWS DEFENDER: Found 'WindowsDefender' key! Its value is " + FileSettingsSave.WindowsDefenderStatus);
-                }
-            }
         }
 
         /* Moved "runAsAdmin" Code to Gist */
